moving bot apis under private folder
