+++ src/private/bot.ts
