+++ test/MicrosoftTeams.spec.ts
@@ -6,7 +6,7 @@ import { TabInstanceParameters, Context, TaskInfo, OpenConversationRequest } fro
 import { TeamInstanceParameters } from "../src/private/interfaces";
 import { TeamType, UserTeamRole, HostClientType, TaskModuleDimension } from "../src/public/constants";
 import { tasks } from "../src/public/tasks";
-import { bot } from "../src/public/bot";
+import { bot } from "../src/private/bot";
 import { conversations } from "../src/private/conversations";
 import { executeDeepLink } from "../src/public/publicAPIs";
 import { frameContexts } from "../src/internal/constants";
